chore: change user services directories structure
