+++ src/domain/user/services/user-password.service.spec.ts
@@ -1,0 +1,3 @@
+describe("", () => {
+  it("", () => expect(1).toBe(1));
+});
