+++ src/domain/user/services/user-password.service.ts
@@ -1,0 +1,4 @@
+import { Injectable } from "@nestjs/common";
+
+@Injectable()
+export class UserPasswordService {}
