+++ src/domain/user/services/user.service.spec.ts
@@ -3,10 +3,10 @@
 
 import { Test, TestingModule } from "@nestjs/testing";
 import { UserService } from "./user.service";
-import { UserRepository } from "./user.repository";
+import { UserRepository } from "../user.repository";
 import { HasherProtocol } from "src/common/hasher/hasher.protocol";
 import { BadRequestException } from "@nestjs/common";
-import { ICreateUser } from "./interfaces/user";
+import { ICreateUser } from "../interfaces/user";
 import { EmailService } from "src/infra/email/email.service";
 import { SanitizeService } from "src/common/sanitize/sanitize.service";
 import { SanitizeProtocol } from "src/common/sanitize/sanitize.protocol";
@@ -18,10 +18,6 @@ const mockUserRepository = {
   findOneById: jest.fn(),
   update: jest.fn(),
   softDelete: jest.fn(),
-  findOneByIdWithCredentials: jest.fn(),
-  findOneByEmailWithCredentials: jest.fn(),
-  saveRefreshTokenAndLastLoginAt: jest.fn(),
-  updateRefreshToken: jest.fn(),
 };
 
 const mockHasherService = {
@@ -254,105 +250,6 @@ describe("UserService", () => {
       expect(userRepositoryMock.findOneById).toHaveBeenCalledWith(
         unexistentUserId,
       );
-
-      expect(result).toBe(null);
-    });
-  });
-
-  describe("findOneByIdWithCredentials", () => {
-    it("should find a user by ID including userCredentials relation", async () => {
-      const storedUser = createMockStoredUser({
-        userCredentials: {
-          passwordHash: "hashed-password",
-          refreshTokenHash: "hashed-token",
-        },
-      });
-
-      mockUserRepository.findOneByIdWithCredentials.mockResolvedValue(
-        storedUser,
-      );
-
-      const userId = "1";
-
-      const result = await userService.findOneByIdWithCredentials(userId);
-      console.log(result);
-
-      expect(
-        userRepositoryMock.findOneByIdWithCredentials,
-      ).toHaveBeenCalledWith(userId);
-
-      expect(result).toMatchObject({
-        id: "1",
-        name: "John Doe",
-        email: "john@email.com",
-        userCredentials: {
-          passwordHash: "hashed-password",
-          refreshTokenHash: "hashed-token",
-        },
-        roles: [{ name: "USER" }],
-      });
-    });
-
-    it("should return null if user don't exist", async () => {
-      mockUserRepository.findOneByIdWithCredentials.mockResolvedValue(null);
-
-      const unexistentUserId = "unexistent-id";
-
-      const result =
-        await userService.findOneByIdWithCredentials(unexistentUserId);
-
-      expect(
-        userRepositoryMock.findOneByIdWithCredentials,
-      ).toHaveBeenCalledWith(unexistentUserId);
-
-      expect(result).toBe(null);
-    });
-  });
-
-  describe("findOneByEmailWithCredentials", () => {
-    it("should find a user by email including userCredentials relation", async () => {
-      const storedUser = createMockStoredUser({
-        userCredentials: {
-          passwordHash: "hashed-password",
-          refreshTokenHash: "hashed-token",
-        },
-      });
-
-      mockUserRepository.findOneByEmailWithCredentials.mockResolvedValue(
-        storedUser,
-      );
-
-      const userEmail = "john@email.com";
-
-      const result = await userService.findOneByEmailWithCredentials(userEmail);
-
-      expect(
-        userRepositoryMock.findOneByEmailWithCredentials,
-      ).toHaveBeenCalledWith(userEmail);
-
-      expect(result).toMatchObject({
-        id: "1",
-        name: "John Doe",
-        email: "john@email.com",
-        userCredentials: {
-          passwordHash: "hashed-password",
-          refreshTokenHash: "hashed-token",
-        },
-        roles: [{ name: "USER" }],
-      });
-    });
-
-    it("should return null if user don't exist", async () => {
-      mockUserRepository.findOneByEmailWithCredentials.mockResolvedValue(null);
-
-      const unexistentUserEmail = "unexistent@email.com";
-
-      const result =
-        await userService.findOneByEmailWithCredentials(unexistentUserEmail);
-
-      expect(
-        userRepositoryMock.findOneByEmailWithCredentials,
-      ).toHaveBeenCalledWith(unexistentUserEmail);
 
       expect(result).toBe(null);
     });
@@ -640,119 +537,6 @@ describe("UserService", () => {
       );
 
       expect(userRepositoryMock.softDelete).not.toHaveBeenCalled();
-    });
-  });
-
-  describe("saveRefreshTokenAndLastLoginAt", () => {
-    it("should save lastLoginAt and refresh token after login", async () => {
-      const userId = "1";
-      const hashedRefreshToken = "hashed-refresh-token";
-      const mockStoredUser = createMockStoredUser({
-        userCredentials: {
-          id: "11",
-          lastLoginAt: new Date(),
-          refreshTokenHash: hashedRefreshToken,
-        },
-      });
-
-      mockUserRepository.saveRefreshTokenAndLastLoginAt.mockResolvedValue(
-        mockStoredUser,
-      );
-
-      const result = await userService.saveRefreshTokenAndLastLoginAt(
-        userId,
-        hashedRefreshToken,
-      );
-
-      expect(
-        userRepositoryMock.saveRefreshTokenAndLastLoginAt,
-      ).toHaveBeenCalledWith(userId, hashedRefreshToken);
-
-      expect(result).not.toHaveProperty("password");
-      expect(result.userCredentials!.lastLoginAt).not.toBeNull();
-
-      expect(result).toMatchObject({
-        id: "1",
-        name: "John Doe",
-        email: "john@email.com",
-        userCredentials: {
-          id: "11",
-          lastLoginAt: expect.any(Date),
-          refreshTokenHash: hashedRefreshToken,
-        },
-        roles: [{ name: "USER" }],
-      });
-    });
-  });
-
-  describe("updateRefreshToken", () => {
-    it("should update refresh token", async () => {
-      const userId = "1";
-      const updatedRefreshTokenHash = "updated-refresh-token-hash";
-      const mockStoredUser = createMockStoredUser({
-        userCredentials: {
-          id: "11",
-          lastLoginAt: new Date(),
-          refreshTokenHash: updatedRefreshTokenHash,
-        },
-      });
-
-      mockUserRepository.updateRefreshToken.mockResolvedValue(mockStoredUser);
-
-      const result = await userService.updateRefreshToken(
-        userId,
-        updatedRefreshTokenHash,
-      );
-
-      expect(userRepositoryMock.updateRefreshToken).toHaveBeenCalledWith(
-        userId,
-        updatedRefreshTokenHash,
-      );
-
-      expect(result).toMatchObject({
-        id: "1",
-        name: "John Doe",
-        email: "john@email.com",
-        userCredentials: {
-          id: "11",
-          lastLoginAt: expect.any(Date),
-          refreshTokenHash: updatedRefreshTokenHash,
-        },
-        roles: [{ name: "USER" }],
-      });
-    });
-
-    it("should invalidate the refresh token", async () => {
-      const userId = "1";
-      const refreshToken = null;
-      const mockStoredUser = createMockStoredUser({
-        userCredentials: {
-          id: "11",
-          lastLoginAt: new Date(),
-          refreshTokenHash: refreshToken,
-        },
-      });
-
-      mockUserRepository.updateRefreshToken.mockResolvedValue(mockStoredUser);
-
-      const result = await userService.updateRefreshToken(userId, refreshToken);
-
-      expect(userRepositoryMock.updateRefreshToken).toHaveBeenCalledWith(
-        userId,
-        refreshToken,
-      );
-
-      expect(result).toMatchObject({
-        id: "1",
-        name: "John Doe",
-        email: "john@email.com",
-        userCredentials: {
-          id: "11",
-          lastLoginAt: expect.any(Date),
-          refreshTokenHash: refreshToken,
-        },
-        roles: [{ name: "USER" }],
-      });
     });
   });
 });
