+++ src/domain/user/user.module.ts
@@ -1,16 +1,23 @@
 import { Module } from "@nestjs/common";
 import { UserControllerV1 } from "./v1/user.controller";
 import { UserControllerV2 } from "./v2/user.controller";
-import { UserService } from "./user.service";
+import { UserService } from "./services/user.service";
 import { UserRepository } from "./user.repository";
 import { PrismaModule } from "src/infra/prisma/prisma.module";
 import { CommonModule } from "src/common/common.module";
 import { EmailModule } from "src/infra/email/email.module";
+import { UserAuthService } from "./services/user-auth.service";
+import { UserPasswordService } from "./services/user-password.service";
 
 @Module({
   imports: [PrismaModule, CommonModule, EmailModule],
   controllers: [UserControllerV1, UserControllerV2],
-  providers: [UserService, UserRepository],
-  exports: [UserService],
+  providers: [
+    UserRepository,
+    UserService,
+    UserAuthService,
+    UserPasswordService,
+  ],
+  exports: [UserService, UserAuthService, UserPasswordService],
 })
 export class UserModule {}
